requirements.txt
python-multipart
fastapi==0.84.0
orjson==3.8.0
pydantic[dotenv]==1.10.2
uvicorn==0.18.3
uvloop==0.17.0; sys_platform == "linux" or sys_platform == "darwin"
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
asyncpg==0.26.0
psycopg2-binary==2.9.3
databases[postgresql]==0.6.1
alembic==1.12.1
passlib
python-jose
aioboto3
aiofiles
